Create package from code
